feat(api-server): implement find-users api
